RealWorldProjects/RedAmon/gvm_scan/requirements.txt
# Requirements for GVM vulnerability scanner container
python-gvm>=24.0.0
python-dotenv>=1.0.0
xmltodict>=0.13.0

